Modified handler to respond only to PR merge
